--- barretenberg/cpp/.clang-format-ignore
@@ -1,1 +1,0 @@
-src/barretenberg/vm2/common/protocol_contract_data.cpp
