+++ barretenberg/cpp/pil/vm2/constants_gen.pil
@@ -15,8 +15,13 @@ namespace constants;
     pol MAX_L2_TO_L1_MSGS_PER_TX = 8;
     pol MAX_PUBLIC_LOGS_PER_TX = 8;
     pol MAX_PACKED_PUBLIC_BYTECODE_SIZE_IN_FIELDS = 3000;
+    pol CANONICAL_AUTH_REGISTRY_ADDRESS = 1;
     pol CONTRACT_INSTANCE_REGISTRY_CONTRACT_ADDRESS = 2;
+    pol CONTRACT_CLASS_REGISTRY_CONTRACT_ADDRESS = 3;
+    pol MULTI_CALL_ENTRYPOINT_ADDRESS = 4;
     pol FEE_JUICE_ADDRESS = 5;
+    pol ROUTER_ADDRESS = 6;
+    pol MAX_PROTOCOL_CONTRACT_ADDRESS = 6;
     pol FEE_JUICE_BALANCES_SLOT = 1;
     pol UPDATED_CLASS_IDS_SLOT = 1;
     pol PUBLIC_LOG_SIZE_IN_FIELDS = 13;
