+++ barretenberg/cpp/src/barretenberg/vm2/common/aztec_constants.hpp
@@ -25,6 +25,7 @@
 #define MULTI_CALL_ENTRYPOINT_ADDRESS 4
 #define FEE_JUICE_ADDRESS 5
 #define ROUTER_ADDRESS 6
+#define MAX_PROTOCOL_CONTRACT_ADDRESS 6
 #define FEE_JUICE_BALANCES_SLOT 1
 #define UPDATED_CLASS_IDS_SLOT 1
 #define PUBLIC_LOG_SIZE_IN_FIELDS 13
