+++ yarn-project/constants/src/constants.gen.ts
@@ -107,6 +107,7 @@ export const CONTRACT_CLASS_REGISTRY_CONTRACT_ADDRESS = 3;
 export const MULTI_CALL_ENTRYPOINT_ADDRESS = 4;
 export const FEE_JUICE_ADDRESS = 5;
 export const ROUTER_ADDRESS = 6;
+export const MAX_PROTOCOL_CONTRACT_ADDRESS = 6;
 export const SIDE_EFFECT_MASKING_ADDRESS = 7;
 export const CONTRACT_CLASS_REGISTRY_BYTECODE_CAPSULE_SLOT = 79025834455612;
 export const FEE_JUICE_BALANCES_SLOT = 1;
