+++ yarn-project/constants/src/scripts/constants.in.ts
@@ -105,6 +105,7 @@ const CPP_CONSTANTS = [
   'L1_TO_L2_MSG_TREE_LEAF_COUNT',
   'PROTOCOL_CONTRACT_TREE_HEIGHT',
   'MAX_PROTOCOL_CONTRACTS',
+  'MAX_PROTOCOL_CONTRACT_ADDRESS',
 ];
 
 const CPP_GENERATORS: string[] = [
@@ -148,8 +149,12 @@ const PIL_CONSTANTS = [
   'L1_TO_L2_MSG_TREE_HEIGHT',
   'UPDATED_CLASS_IDS_SLOT',
   'UPDATES_DELAYED_PUBLIC_MUTABLE_VALUES_LEN',
+  'CANONICAL_AUTH_REGISTRY_ADDRESS',
   'CONTRACT_INSTANCE_REGISTRY_CONTRACT_ADDRESS',
+  'CONTRACT_CLASS_REGISTRY_CONTRACT_ADDRESS',
+  'MULTI_CALL_ENTRYPOINT_ADDRESS',
   'FEE_JUICE_ADDRESS',
+  'ROUTER_ADDRESS',
   'FEE_JUICE_BALANCES_SLOT',
   'TIMESTAMP_OF_CHANGE_BIT_SIZE',
   'UPDATES_DELAYED_PUBLIC_MUTABLE_METADATA_BIT_SIZE',
@@ -279,6 +284,7 @@ const PIL_CONSTANTS = [
   'NOTE_HASH_TREE_LEAF_COUNT',
   'L1_TO_L2_MSG_TREE_LEAF_COUNT',
   'PROTOCOL_CONTRACT_TREE_HEIGHT',
+  'MAX_PROTOCOL_CONTRACT_ADDRESS',
 ];
 
 const PIL_GENERATORS: string[] = [
