+++ yarn-project/simulator/artifacts/avm_minimal_inputs.json
@@ -8,23 +8,45 @@
       "timestamp": "99833",
       "coinbase": "0x0000000000000000000000000000000000000000",
       "feeRecipient": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "gasFees": { "feePerDaGas": "2", "feePerL2Gas": "3" }
+      "gasFees": {
+        "feePerDaGas": "2",
+        "feePerL2Gas": "3"
+      }
     },
     "tx": {
       "hash": "0x163c46f0521b97ce6899de56363a018548f78d331df7561a503fd72da72be6fe",
       "gasSettings": {
-        "gasLimits": { "daGas": 6291456, "l2Gas": 6000000 },
-        "teardownGasLimits": { "daGas": 0, "l2Gas": 0 },
-        "maxFeesPerGas": { "feePerDaGas": "10", "feePerL2Gas": "10" },
-        "maxPriorityFeesPerGas": { "feePerDaGas": "0", "feePerL2Gas": "0" }
+        "gasLimits": {
+          "daGas": 6291456,
+          "l2Gas": 6000000
+        },
+        "teardownGasLimits": {
+          "daGas": 0,
+          "l2Gas": 0
+        },
+        "maxFeesPerGas": {
+          "feePerDaGas": "10",
+          "feePerL2Gas": "10"
+        },
+        "maxPriorityFeesPerGas": {
+          "feePerDaGas": "0",
+          "feePerL2Gas": "0"
+        }
       },
-      "effectiveGasFees": { "feePerDaGas": "2", "feePerL2Gas": "3" },
+      "effectiveGasFees": {
+        "feePerDaGas": "2",
+        "feePerL2Gas": "3"
+      },
       "nonRevertibleAccumulatedData": {
         "noteHashes": [],
         "nullifiers": ["0x00000000000000000000000000000000000000000000000000000000000668a0"],
         "l2ToL1Messages": []
       },
-      "revertibleAccumulatedData": { "noteHashes": [], "nullifiers": [], "l2ToL1Messages": [] },
+      "revertibleAccumulatedData": {
+        "noteHashes": [],
+        "nullifiers": [],
+        "l2ToL1Messages": []
+      },
       "setupEnqueuedCalls": [],
       "appLogicEnqueuedCalls": [
         {
@@ -38,7 +60,10 @@
         }
       ],
       "teardownEnqueuedCall": null,
-      "gasUsedByPrivate": { "daGas": 0, "l2Gas": 0 },
+      "gasUsedByPrivate": {
+        "daGas": 0,
+        "l2Gas": 0
+      },
       "feePayer": "0x000000000000000000000000000000000000000000000000000000000000002a"
     },
     "protocolContractDerivedAddresses": [
@@ -624,7 +649,9 @@
         },
         "index": "127",
         "leafPreimage": {
-          "leaf": { "nullifier": "0x000000000000000000000000000000000000000000000000000000000000007f" },
+          "leaf": {
+            "nullifier": "0x000000000000000000000000000000000000000000000000000000000000007f"
+          },
           "nextKey": "0x27252dad7f80281796fbc24345842e0c4322466d6997fed605a9c70d5a5fe4a4",
           "nextIndex": "128"
         }
@@ -636,7 +663,9 @@
         },
         "index": "128",
         "leafPreimage": {
-          "leaf": { "nullifier": "0x27252dad7f80281796fbc24345842e0c4322466d6997fed605a9c70d5a5fe4a4" },
+          "leaf": {
+            "nullifier": "0x27252dad7f80281796fbc24345842e0c4322466d6997fed605a9c70d5a5fe4a4"
+          },
           "nextKey": "0x0000000000000000000000000000000000000000000000000000000000000000",
           "nextIndex": "0"
         }
@@ -777,10 +806,14 @@
           "nextAvailableLeafIndex": 130
         },
         "treeId": 0,
-        "leaf": { "nullifier": "0x00000000000000000000000000000000000000000000000000000000000668a0" },
+        "leaf": {
+          "nullifier": "0x00000000000000000000000000000000000000000000000000000000000668a0"
+        },
         "lowLeavesWitnessData": {
           "leaf": {
-            "leaf": { "nullifier": "0x000000000000000000000000000000000000000000000000000000000000007f" },
+            "leaf": {
+              "nullifier": "0x000000000000000000000000000000000000000000000000000000000000007f"
+            },
             "nextKey": "0x27252dad7f80281796fbc24345842e0c4322466d6997fed605a9c70d5a5fe4a4",
             "nextIndex": "128"
           },
@@ -830,7 +863,9 @@
         },
         "insertionWitnessData": {
           "leaf": {
-            "leaf": { "nullifier": "0x0000000000000000000000000000000000000000000000000000000000000000" },
+            "leaf": {
+              "nullifier": "0x0000000000000000000000000000000000000000000000000000000000000000"
+            },
             "nextKey": "0x0000000000000000000000000000000000000000000000000000000000000000",
             "nextIndex": "0"
           },
@@ -881,8 +916,20 @@
       }
     ],
     "appendLeavesHints": [],
-    "createCheckpointHints": [{ "actionCounter": 0, "oldCheckpointId": 0, "newCheckpointId": 1 }],
-    "commitCheckpointHints": [{ "actionCounter": 1, "oldCheckpointId": 1, "newCheckpointId": 0 }],
+    "createCheckpointHints": [
+      {
+        "actionCounter": 0,
+        "oldCheckpointId": 0,
+        "newCheckpointId": 1
+      }
+    ],
+    "commitCheckpointHints": [
+      {
+        "actionCounter": 1,
+        "oldCheckpointId": 1,
+        "newCheckpointId": 0
+      }
+    ],
     "revertCheckpointHints": []
   },
   "publicInputs": {
@@ -894,7 +941,10 @@
       "timestamp": "99833",
       "coinbase": "0x0000000000000000000000000000000000000000",
       "feeRecipient": "0x0000000000000000000000000000000000000000000000000000000000000000",
-      "gasFees": { "feePerDaGas": "2", "feePerL2Gas": "3" }
+      "gasFees": {
+        "feePerDaGas": "2",
+        "feePerL2Gas": "3"
+      }
     },
     "startTreeSnapshots": {
       "l1ToL2MessageTree": {
@@ -914,16 +964,38 @@
         "nextAvailableLeafIndex": 129
       }
     },
-    "startGasUsed": { "daGas": 0, "l2Gas": 0 },
-    "gasSettings": {
-      "gasLimits": { "daGas": 6291456, "l2Gas": 6000000 },
-      "teardownGasLimits": { "daGas": 0, "l2Gas": 0 },
-      "maxFeesPerGas": { "feePerDaGas": "10", "feePerL2Gas": "10" },
-      "maxPriorityFeesPerGas": { "feePerDaGas": "0", "feePerL2Gas": "0" }
+    "startGasUsed": {
+      "daGas": 0,
+      "l2Gas": 0
     },
-    "effectiveGasFees": { "feePerDaGas": "2", "feePerL2Gas": "3" },
+    "gasSettings": {
+      "gasLimits": {
+        "daGas": 6291456,
+        "l2Gas": 6000000
+      },
+      "teardownGasLimits": {
+        "daGas": 0,
+        "l2Gas": 0
+      },
+      "maxFeesPerGas": {
+        "feePerDaGas": "10",
+        "feePerL2Gas": "10"
+      },
+      "maxPriorityFeesPerGas": {
+        "feePerDaGas": "0",
+        "feePerL2Gas": "0"
+      }
+    },
+    "effectiveGasFees": {
+      "feePerDaGas": "2",
+      "feePerL2Gas": "3"
+    },
     "feePayer": "0x000000000000000000000000000000000000000000000000000000000000002a",
-    "publicCallRequestArrayLengths": { "setupCalls": 0, "appLogicCalls": 1, "teardownCall": false },
+    "publicCallRequestArrayLengths": {
+      "setupCalls": 0,
+      "appLogicCalls": 1,
+      "teardownCall": false
+    },
     "publicSetupCallRequests": [
       {
         "msgSender": "0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -1318,8 +1390,16 @@
       "isStaticCall": false,
       "calldataHash": "0x0000000000000000000000000000000000000000000000000000000000000000"
     },
-    "previousNonRevertibleAccumulatedDataArrayLengths": { "noteHashes": 0, "nullifiers": 1, "l2ToL1Msgs": 0 },
-    "previousRevertibleAccumulatedDataArrayLengths": { "noteHashes": 0, "nullifiers": 0, "l2ToL1Msgs": 0 },
+    "previousNonRevertibleAccumulatedDataArrayLengths": {
+      "noteHashes": 0,
+      "nullifiers": 1,
+      "l2ToL1Msgs": 0
+    },
+    "previousRevertibleAccumulatedDataArrayLengths": {
+      "noteHashes": 0,
+      "nullifiers": 0,
+      "l2ToL1Msgs": 0
+    },
     "previousNonRevertibleAccumulatedData": {
       "noteHashes": [
         "0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -1722,7 +1802,10 @@
         "nextAvailableLeafIndex": 129
       }
     },
-    "endGasUsed": { "daGas": 0, "l2Gas": 75 },
+    "endGasUsed": {
+      "daGas": 0,
+      "l2Gas": 75
+    },
     "accumulatedDataArrayLengths": {
       "noteHashes": 0,
       "nullifiers": 1,
